CallbackWrapper: int->size_t

+++ include/callbacks/callback.h
@@ -10,10 +10,10 @@ class Callback {
   Callback() {}
   virtual ~Callback() {}
 
-  virtual void OnEpochBegin(const int index) = 0;
-  virtual void OnEpochEnd(const int index) = 0;
-  virtual void OnBatchBegin(const int index) = 0;
-  virtual void OnBatchEnd(const int index, const float loss) = 0;
+  virtual void OnEpochBegin(const size_t index) = 0;
+  virtual void OnEpochEnd(const size_t index) = 0;
+  virtual void OnBatchBegin(const size_t index) = 0;
+  virtual void OnBatchEnd(const size_t index, const float loss) = 0;
 
   DISABLE_COPY_AND_ASSIGN(Callback);
 };
