+++ src/callbacks/callback_wrapper.cc
@@ -4,28 +4,28 @@
 
 namespace blitz {
 
-void CallbackWrapper::OnEpochBegin(const int index) {
+void CallbackWrapper::OnEpochBegin(const size_t index) {
   for (list<shared_ptr<Callback> >::iterator it = callbacks_.begin();
       it != callbacks_.end(); ++it) {
     (*it)->OnEpochBegin(index);
   }
 }
 
-void CallbackWrapper::OnEpochEnd(const int index) {
+void CallbackWrapper::OnEpochEnd(const size_t index) {
   for (list<shared_ptr<Callback> >::iterator it = callbacks_.begin();
       it != callbacks_.end(); ++it) {
     (*it)->OnEpochEnd(index);
   }
 }
 
-void CallbackWrapper::OnBatchBegin(const int index) {
+void CallbackWrapper::OnBatchBegin(const size_t index) {
   for (list<shared_ptr<Callback> >::iterator it = callbacks_.begin();
       it != callbacks_.end(); ++it) {
     (*it)->OnBatchBegin(index);
   }
 }
 
-void CallbackWrapper::OnBatchEnd(const int index, const float loss) {
+void CallbackWrapper::OnBatchEnd(const size_t index, const float loss) {
   for (list<shared_ptr<Callback> >::iterator it = callbacks_.begin();
       it != callbacks_.end(); ++it) {
     (*it)->OnBatchEnd(index, loss);
